requirements.txt
PySide6-Fluent-Widgets
google_trans_new
requests
pyperclip
tqdm
